packet

+++ uLab/Assets/Scripts/Network/Protocal.cs
@@ -4,8 +4,8 @@ namespace Locke
 	public class Protocal
 	{
 		///BUILD TABLE
-		public const int Connect = 101;     //连接服务器
-		public const int Exception = 102;     //异常掉线
-		public const int Disconnect = 103;     //正常断线   
+		public const ushort Connect = 101;     //连接服务器
+		public const ushort Exception = 102;     //异常掉线
+		public const ushort Disconnect = 103;     //正常断线   
 	}
 }
